+++ waspc/cli/src/Wasp/Cli/Command/Start/Db.hs
@@ -23,8 +23,9 @@ import Wasp.Cli.Command.Compile (analyze)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
 import Wasp.Cli.Command.Require.WaspSpecAvailable (WaspSpecAvailable (WaspSpecAvailable))
-import Wasp.Cli.Port (findFirstFreeLocalPortInRange)
+import Wasp.Cli.Port (resolvePort)
 import Wasp.Cli.Util.Parser (withArguments)
+import Wasp.Cli.Util.PortArgument (portOption)
 import Wasp.Db.Postgres (defaultPostgresDockerImageSpec, defaultPostgresPort)
 import qualified Wasp.Message as Msg
 import Wasp.Project.Common (WaspProjectDir)
@@ -53,6 +54,7 @@ start = withArguments "wasp start db" startDbArgsParser $ \args -> do
       startPostgresDevDb
         waspProjectDir
         appName
+        (dbPort args)
         (dbImage args)
         (dbVolumeMountPath args)
   where
@@ -63,7 +65,8 @@ start = withArguments "wasp start db" startDbArgsParser $ \args -> do
 startDbArgsParser :: Opt.Parser StartDbArgs
 startDbArgsParser =
   StartDbArgs
-    <$> Opt.strOption
+    <$> portOption "db-port" ("Port to run the dev database on (default: " ++ show defaultPostgresPort ++ ")")
+    <*> Opt.strOption
       ( Opt.long "db-image"
           <> Opt.metavar "IMAGE"
           <> Opt.help "Docker image to use for the database"
@@ -79,7 +82,8 @@ startDbArgsParser =
       )
 
 data StartDbArgs = StartDbArgs
-  { dbImage :: DockerImageName,
+  { dbPort :: Maybe PortNumber,
+    dbImage :: DockerImageName,
     dbVolumeMountPath :: DockerVolumeMountPath
   }
 
@@ -117,28 +121,24 @@ throwIfCustomDbAlreadyInUse spec = do
     throwCustomDbAlreadyInUseError msg =
       E.throwError $ CommandError "You are using custom database already" msg
 
-startPostgresDevDb :: Path' Abs (Dir WaspProjectDir) -> String -> DockerImageName -> DockerVolumeMountPath -> Command ()
-startPostgresDevDb waspProjectDir appName dbDockerImage dbDockerVolumeMountPath = do
+startPostgresDevDb :: Path' Abs (Dir WaspProjectDir) -> String -> Maybe PortNumber -> DockerImageName -> DockerVolumeMountPath -> Command ()
+startPostgresDevDb waspProjectDir appName requestedDbPort dbDockerImage dbDockerVolumeMountPath = do
   throwIfExeIsNotAvailable
     "docker"
     "To run PostgreSQL dev database, Wasp needs `docker` installed and in PATH."
 
   liftIO (Dev.Postgres.discoverProjectsRunningDevDb waspProjectDir appName) >>= \case
     Just runningDb -> noteDbIsAlreadyRunningAndExit runningDb
-    Nothing -> startDbOnPort =<< findFreeDevDbPort
+    Nothing -> startDbOnPort =<< resolveDevDbPort
   where
-    findFreeDevDbPort :: Command PortNumber
-    findFreeDevDbPort =
-      liftIO
-        ( findFirstFreeLocalPortInRange
-            defaultPostgresPort
-            []
-            "Free at least one of those ports by exiting the program listening on it."
-        )
-        >>= either throwNoFreePortError return
-
-    throwNoFreePortError :: String -> Command a
-    throwNoFreePortError = E.throwError . CommandError "No free port"
+    resolveDevDbPort :: Command PortNumber
+    resolveDevDbPort =
+      resolvePort
+        requestedDbPort
+        defaultPostgresPort
+        []
+        "Choose a different port with --db-port, or free up this one."
+        "Free at least one of those ports by exiting the program listening on it, or choose one yourself with --db-port."
 
     noteDbIsAlreadyRunningAndExit :: Dev.Postgres.DevDbSpec -> Command ()
     noteDbIsAlreadyRunningAndExit devDbSpec = do